+++ lib/strong_migrations/checker.rb
@@ -101,13 +101,33 @@ Then add the NOT NULL constraint."
             raise_error :add_column_json
           end
         when :change_column
-          table, column, type = args
+          table, column, type, options = args
+          options ||= {}
 
           safe = false
-          # assume Postgres 9.1+ since previous versions are EOL
-          if postgresql? && type.to_s == "text"
+          # assume Postgres 9.5+ since previous versions are EOL
+          if postgresql?
             found_column = connection.columns(table).find { |c| c.name.to_s == column.to_s }
-            safe = found_column && found_column.type == :string
+            if found_column
+              case type.to_s
+              when "text"
+                safe = found_column.type == :string
+              when "numeric", "decimal"
+                # numeric and decimal are equivalent and can be used interchangably
+                safe = [:numeric, :decimal].include?(found_column.type) &&
+                  (
+                    (
+                      # unconstrained
+                      !options[:precision] && !options[:scale]
+                    ) || (
+                      # increased precision, same scale
+                      options[:precision] && found_column.precision &&
+                      options[:precision] >= found_column.precision &&
+                      options[:scale] == found_column.scale
+                    )
+                  )
+              end
+            end
           end
           raise_error :change_column unless safe
         when :create_table